Move kubernetes-client-internal code to proper packages
Fixes: #44730
Fixes: #44731
