+++ extensions/kubernetes-client/deployment-internal/src/main/java/io/quarkus/kubernetes/client/deployment/internal/KubernetesClientBuildStep.java
@@ -1,11 +1,11 @@
-package io.quarkus.kubernetes.client.deployment;
+package io.quarkus.kubernetes.client.deployment.internal;
 
-import static io.quarkus.kubernetes.client.runtime.KubernetesClientUtils.createConfig;
+import static io.quarkus.kubernetes.client.runtime.internal.KubernetesClientUtils.createConfig;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.QuarkusBuildCloseablesBuildItem;
-import io.quarkus.kubernetes.client.runtime.KubernetesClientBuildConfig;
-import io.quarkus.kubernetes.client.runtime.QuarkusHttpClientFactory;
+import io.quarkus.kubernetes.client.runtime.internal.KubernetesClientBuildConfig;
+import io.quarkus.kubernetes.client.runtime.internal.QuarkusHttpClientFactory;
 import io.quarkus.kubernetes.client.spi.KubernetesClientBuildItem;
 
 public class KubernetesClientBuildStep {
