+++ extensions/kubernetes-client/runtime-internal/src/main/resources/META-INF/services/io.fabric8.kubernetes.client.http.HttpClient$Factory
@@ -1,1 +1,1 @@
-io.quarkus.kubernetes.client.runtime.QuarkusHttpClientFactory
+io.quarkus.kubernetes.client.runtime.internal.QuarkusHttpClientFactory
