+++ extensions/kubernetes-client/runtime/src/main/java/io/quarkus/kubernetes/client/KubernetesConfigCustomizer.java
@@ -5,9 +5,9 @@ import java.util.List;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.utils.KubernetesSerialization;
-import io.quarkus.kubernetes.client.runtime.KubernetesClientBuildConfig;
 import io.quarkus.kubernetes.client.runtime.KubernetesClientProducer;
 import io.quarkus.kubernetes.client.runtime.KubernetesConfigProducer;
+import io.quarkus.kubernetes.client.runtime.internal.KubernetesClientBuildConfig;
 
 /**
  * Meant to be implemented by a CDI bean that provided arbitrary customization for the default {@link Config} created by
